# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: 鲸云AI教育平台 API
  version: 0.1.0
  description: |
    W0-2 契约 · MVP 口径(对应《MVP 裁剪手册》)。统一响应包 {code,message,data};
    分页 {items,total};鉴权 Bearer JWT,payload 含 {uid,orgId,role}。
    角色约束写在各接口 description 的 [角色] 标注中。
servers: [{ url: /api/v1 }]
security: [{ bearer: [] }]
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    page: { name: page, in: query, schema: { type: integer, default: 1, minimum: 1 } }
    size: { name: size, in: query, schema: { type: integer, default: 20, maximum: 50 } }
    keyword: { name: keyword, in: query, schema: { type: string } }
    idPath: { name: id, in: path, required: true, schema: { type: integer } }
  responses:
    Err:
      description: 业务/鉴权错误(401 未登录 403 越权 404 不存在或跨租户)
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrResp' } } }
  schemas:
    ErrResp:
      type: object
      required: [code, message]
      properties: { code: { type: integer }, message: { type: string }, detail: {} }
    OkVoid:
      type: object
      required: [code, message, data]
      properties: { code: { type: integer, enum: [0] }, message: { type: string }, data: { type: 'null' } }
    # ---- 枚举 ----
...
    ResourceType: { type: string, enum: [ppt, pdf, video, interactive, image] }
    QuestionType: { type: string, enum: [single, multi, blank, solution] }
...
    PaperType: { type: string, enum: [homework, exam, practice] }
...
    GraphType: { type: string, enum: [curriculum_knowledge, problem_solving_ability, problem_solving_strategy] }
...
    Resource:
      type: object
      required: [id, type, name, ossKey, size, meta, usedByLessons, kpNodeId, kpNodeName, createdAt]
      properties:
        id: { type: integer }
        type: { $ref: '#/components/schemas/ResourceType' }
        name: { type: string }
        ossKey: { type: string }
        size: { type: integer }
        meta: { type: object, additionalProperties: true }
        usedByLessons:
          type: array
          items:
            type: object
            required: [lessonId, lessonTitle]
            properties: { lessonId: { type: integer }, lessonTitle: { type: string } }
        kpNodeId: { oneOf: [{ type: 'null' }, { type: integer }], description: 归档知识点(只读) }
        kpNodeName: { oneOf: [{ type: 'null' }, { type: string }], description: 知识点名称(只读) }
        createdAt: { type: string, format: date-time }
...
    RubricStep:
      type: object
      required: [step, desc, score]
      properties: { step: { type: integer }, desc: { type: string }, score: { type: number } }
...
    Paper:
      type: object
      required: [id, name, type, totalScore, status, questions]
      properties:
        id: { type: integer }
        name: { type: string }
        type: { $ref: '#/components/schemas/PaperType' }
        totalScore: { type: number }
        status: { type: string }
        questions:
          type: array
          items:
            type: object
            required: [seq, questionId, score, type, stemLatex]
            properties:
              seq: { type: integer }
              questionId: { type: integer }
              score: { type: number }
              type: { $ref: '#/components/schemas/QuestionType' }
              stemLatex: { type: string }
    PaperInput:
      type: object
      required: [name, type, questions]
      properties:
        name: { type: string }
        type: { $ref: '#/components/schemas/PaperType' }
        questions:
          type: array
          minItems: 1
          items:
            type: object
            required: [questionId, score]
            properties: { questionId: { type: integer }, score: { type: number } }
...
    GradingAnswerBrief:
      type: object
      required: [answerId, studentId, studentName, questionId, seq, status, aiScore, finalScore]
      properties:
        answerId: { type: integer }
        studentId: { type: integer }
        studentName: { type: string }
        questionId: { type: integer }
        seq: { type: integer }
        status: { type: string, enum: [pending, graded] }
        aiScore: { oneOf: [{ type: 'null' }, { type: number }] }
        finalScore: { oneOf: [{ type: 'null' }, { type: number }] }
...
    GradingItem:
      type: object
      required: [answerId, studentId, studentName, questionId, stemLatex, rubric, photoUrl, textResponse, aiScore, aiSteps, aiErrorTags, finalScore, comment]
      properties:
        answerId: { type: integer }
        studentId: { type: integer }
        studentName: { type: string }
        questionId: { type: integer }
        stemLatex: { type: string }
        rubric: { type: array, items: { $ref: '#/components/schemas/RubricStep' } }
        photoUrl: { oneOf: [{ type: 'null' }, { type: string }], description: 签名URL,短时效 }
        textResponse: { oneOf: [{ type: 'null' }, { type: string }] }
        aiScore: { oneOf: [{ type: 'null' }, { type: number }] }
        aiSteps:
          type: array
          items:
            type: object
            required: [step, ok]
            properties: { step: { type: integer }, ok: { type: boolean }, comment: { type: string } }
        aiErrorTags: { type: array, items: { type: string } }
        finalScore: { oneOf: [{ type: 'null' }, { type: number }] }
        comment: { oneOf: [{ type: 'null' }, { type: string }] }
...
    MasteryItem:
      type: object
      required: [nodeId, nodeName, graphType, mastery, sampleCount]
      properties:
        nodeId: { type: integer }
        nodeName: { type: string }
        graphType: { $ref: '#/components/schemas/GraphType' }
        mastery: { type: integer, minimum: 0, maximum: 100 }
        sampleCount: { type: integer }
...
    AiDiagnosis:
      type: object
      description: AI 学情诊断结果:整体诊断文字 + 薄弱知识点列表
      required: [summary]
      properties:
        summary: { type: string, description: 整体诊断文字 }
        weakPoints:
          type: array
          description: 薄弱知识点列表
          items:
            type: object
            required: [kpName, reason]
            properties: { kpName: { type: string }, reason: { type: string } }
        generatedAt: { type: string, format: date-time }
paths:
...
  /resources:
    get:
      summary: 资源列表(含被引用讲次) [teacher]
      parameters: [{$ref: '#/components/parameters/page'}, {$ref: '#/components/parameters/size'}, {$ref: '#/components/parameters/keyword'}, {name: type, in: query, schema: {$ref: '#/components/schemas/ResourceType'}}]
      responses:
        '200': {description: ok, content: {application/json: {schema: {type: object, required: [code, message, data], properties: {code: {type: integer}, message: {type: string}, data: {type: object, required: [items, total], properties: {items: {type: array, items: {$ref: '#/components/schemas/Resource'}}, total: {type: integer}}}}}}}}
        default: {$ref: '#/components/responses/Err'}
    post:
      summary: 登记已直传的资源 [teacher]
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [type, name, ossKey, size], properties: {type: {$ref: '#/components/schemas/ResourceType'}, name: {type: string}, ossKey: {type: string}, size: {type: integer}, meta: {type: object, additionalProperties: true}, kpNodeId: {oneOf: [{type: 'null'}, {type: integer}], description: '可选:按知识点归档'}}}}}}
      responses:
        '200': {description: ok, content: {application/json: {schema: {type: object, required: [code, message, data], properties: {code: {type: integer}, message: {type: string}, data: {$ref: '#/components/schemas/Resource'}}}}}}
        default: {$ref: '#/components/responses/Err'}
  /resources/{id}:
    put:
      summary: 重命名/更新元信息 [teacher]
      parameters: [{$ref: '#/components/parameters/idPath'}]
      requestBody: {required: true, content: {application/json: {schema: {type: object, properties: {name: {type: string}, meta: {type: object, additionalProperties: true}, kpNodeId: {oneOf: [{type: 'null'}, {type: integer}], description: '可选:按知识点归档,null 清除'}}}}}}
      responses:
        '200': {description: ok, content: {application/json: {schema: {$ref: '#/components/schemas/OkVoid'}}}}
        default: {$ref: '#/components/responses/Err'}
    delete:
      summary: 删除资源(被讲次引用时禁止) [teacher]
      parameters: [{$ref: '#/components/parameters/idPath'}]
      responses:
        '200': {description: ok, content: {application/json: {schema: {$ref: '#/components/schemas/OkVoid'}}}}
        default: {$ref: '#/components/responses/Err'}

  # ================= 教师 · 课程/讲次/编排/组卷/发布 =================
...
  /papers:
    get:
      summary: 试卷列表 [teacher]
      parameters: [{$ref: '#/components/parameters/page'}, {$ref: '#/components/parameters/size'}, {name: type, in: query, schema: {$ref: '#/components/schemas/PaperType'}}]
      responses:
        '200': {description: ok, content: {application/json: {schema: {type: object, required: [code, message, data], properties: {code: {type: integer}, message: {type: string}, data: {type: object, required: [items, total], properties: {items: {type: array, items: {$ref: '#/components/schemas/Paper'}}, total: {type: integer}}}}}}}}
        default: {$ref: '#/components/responses/Err'}
    post:
      summary: 创建试卷(手动挑题) [teacher]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/PaperInput'}}}}
      responses:
        '200': {description: ok, content: {application/json: {schema: {type: object, required: [code, message, data], properties: {code: {type: integer}, message: {type: string}, data: {$ref: '#/components/schemas/Paper'}}}}}}
        default: {$ref: '#/components/responses/Err'}
  /papers/{id}:
    get:
      summary: 试卷详情 [teacher]
      parameters: [{$ref: '#/components/parameters/idPath'}]
      responses:
        '200': {description: ok, content: {application/json: {schema: {type: object, required: [code, message, data], properties: {code: {type: integer}, message: {type: string}, data: {$ref: '#/components/schemas/Paper'}}}}}}
        default: {$ref: '#/components/responses/Err'}
    put:
      summary: 增删题/调分(已下发的禁止) [teacher]
      parameters: [{$ref: '#/components/parameters/idPath'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/PaperInput'}}}}
      responses:
        '200': {description: ok, content: {application/json: {schema: {$ref: '#/components/schemas/OkVoid'}}}}
        default: {$ref: '#/components/responses/Err'}
...
  /grading/pending:
    get:
      summary: 待复核列表(按作业聚合) [teacher]
      responses:
        '200': {description: ok, content: {application/json: {schema: {type: object, required: [code, message, data], properties: {code: {type: integer}, message: {type: string}, data: {type: array, items: {type: object, required: [assignmentId, paperName, pendingCount, aiAvgScore], properties: {assignmentId: {type: integer}, paperName: {type: string}, pendingCount: {type: integer}, aiAvgScore: {oneOf: [{type: 'null'}, {type: number}]}}}}}}}}}
        default: {$ref: '#/components/responses/Err'}
  /grading/assignments/{id}/answers:
    get:
      summary: 某作业逐题作答名单(待复核/已复核,供教师切换) [teacher]
      parameters: [{$ref: '#/components/parameters/idPath'}, {name: status, in: query, schema: {type: string, enum: [pending, graded]}}]
      responses:
        '200': {description: ok, content: {application/json: {schema: {type: object, required: [code, message, data], properties: {code: {type: integer}, message: {type: string}, data: {type: array, items: {$ref: '#/components/schemas/GradingAnswerBrief'}}}}}}}
        default: {$ref: '#/components/responses/Err'}
  /grading/answers/{id}:
    get:
      summary: 单份复核详情(原稿+AI预批+rubric) [teacher]
      parameters: [{$ref: '#/components/parameters/idPath'}]
      responses:
        '200': {description: ok, content: {application/json: {schema: {type: object, required: [code, message, data], properties: {code: {type: integer}, message: {type: string}, data: {$ref: '#/components/schemas/GradingItem'}}}}}}
        default: {$ref: '#/components/responses/Err'}
  /grading/answers/{id}/review:
    put:
      summary: 确认得分与评语 [teacher]
      parameters: [{$ref: '#/components/parameters/idPath'}]
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [finalScore], properties: {finalScore: {type: number}, comment: {type: string}}}}}}
      responses:
        '200': {description: ok, content: {application/json: {schema: {$ref: '#/components/schemas/OkVoid'}}}}
        default: {$ref: '#/components/responses/Err'}
  /grading/assignments/{id}/finalize:
    post:
      summary: 出分(成绩对学生可见,触发错题/掌握度任务) [teacher]
      parameters: [{$ref: '#/components/parameters/idPath'}]
      responses:
        '200': {description: ok, content: {application/json: {schema: {$ref: '#/components/schemas/OkVoid'}}}}
        default: {$ref: '#/components/responses/Err'}

  # ================= 学生 =================
...
  /analytics/courses/{id}/mastery:
    get:
      summary: 课程知识点掌握热力 [teacher]
      parameters: [{$ref: '#/components/parameters/idPath'}]
      responses:
        '200': {description: ok, content: {application/json: {schema: {type: object, required: [code, message, data], properties: {code: {type: integer}, message: {type: string}, data: {type: array, items: {type: object, required: [nodeId, nodeName, avgMastery, studentCount], properties: {nodeId: {type: integer}, nodeName: {type: string}, avgMastery: {type: integer}, studentCount: {type: integer}}}}}}}}}
        default: {$ref: '#/components/responses/Err'}
  /analytics/courses/{id}/attention:
    get:
      summary: 重点关注学生 [teacher]
      parameters: [{$ref: '#/components/parameters/idPath'}]
      responses:
        '200': {description: ok, content: {application/json: {schema: {type: object, required: [code, message, data], properties: {code: {type: integer}, message: {type: string}, data: {type: array, items: {type: object, required: [studentId, name, reason], properties: {studentId: {type: integer}, name: {type: string}, reason: {type: string}}}}}}}}}
        default: {$ref: '#/components/responses/Err'}
  /analytics/students/{id}:
    get:
      summary: 单生 30 天报告 [teacher/admin]
      parameters: [{$ref: '#/components/parameters/idPath'}]
      responses:
        '200': {description: ok, content: {application/json: {schema: {type: object, required: [code, message, data], properties: {code: {type: integer}, message: {type: string}, data: {type: object, required: [mastery, wrongOpenCount, attempts30d], properties: {mastery: {type: array, items: {$ref: '#/components/schemas/MasteryItem'}}, wrongOpenCount: {type: integer}, attempts30d: {type: integer}}}}}}}}
        default: {$ref: '#/components/responses/Err'}
  /analytics/students/{id}/diagnose:
    post:
      summary: AI 学情诊断 [teacher/admin]
      parameters: [{$ref: '#/components/parameters/idPath'}]
      responses:
        '200': {description: ok, content: {application/json: {schema: {type: object, required: [code, message, data], properties: {code: {type: integer}, message: {type: string}, data: {$ref: '#/components/schemas/AiDiagnosis'}}}}}}
        default: {$ref: '#/components/responses/Err'}

  # ================= AI =================
...